move dns outside http and add url root mod
