+++ src/dns/mod.rs
@@ -1,0 +1,3 @@
+//! DNS module for Rama.
+
+pub mod layer;
