+++ src/http/layer/mod.rs
@@ -8,7 +8,6 @@
 //! - [`auth`]: A layer that can be used to authenticate requests, branching
 //!   in case the request is not authenticated (read: rejected).
 //! - [`cors`]: A layer that can be used to add CORS headers to the response.
-//! - [`dns`]: A layer that can be used to resolve the hostname of the request.
 //!
 //! Most layers are implemented as a [`Service`], and then wrapped in a [`Layer`].
 //! This is done to allow the layer to be used as a service, and to allow it to be
@@ -22,7 +21,6 @@ pub mod body_limit;
 pub mod catch_panic;
 pub mod classify;
 pub mod cors;
-pub mod dns;
 pub mod header_config;
 pub mod map_request_body;
 pub mod map_response_body;
