+++ src/url/mod.rs
@@ -1,0 +1,6 @@
+//! re-export of `url` crate + extras on top of it
+
+pub use url::*;
+
+mod scheme;
+pub use scheme::Scheme;
